Adaptive Simulation Bounds
Simulation bounds now change depending on the farthest body.
This should only be used in systems with orbits, as the gravitational singularity can create issues as there is no cap for the size, as that would be hard to implement because of all simulations having a chance of being at a different scale.